packages/hap-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bluescroll-hap"
version = "0.1.0"
description = "Official HAP (Human Application Protocol) SDK for Python"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "BlueScroll Inc." }
]
keywords = [
    "hap",
    "human-application-protocol",
    "verification",
    "job-applications",
    "ed25519",
    "jws"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/BlueScroll/hap"
Repository = "https://github.com/BlueScroll/hap.git"
Documentation = "https://github.com/BlueScroll/hap#readme"

[tool.hatch.build.targets.wheel]
packages = ["hap"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
